agent: change type of snmp_enableauthentraps to long
Type of snmp_enableauthentraps variable is changed from int to long.
This fixes the bug with writing to snmpEnableAuthenTraps if server is
compiled for 64bit (snmpset reports wrong length). This bug was
introduced a long time ago (by faeecd0 commit).

Actually this is a 'less effort' type of fix but I wonder why we need a
long to store an INTEGER in the first place?

+++ agent/mibgroup/mibII/snmp_mib.c
@@ -33,7 +33,7 @@
 	 *
 	 *********************/
 
-extern int      snmp_enableauthentraps;
+extern long     snmp_enableauthentraps;
 extern int      snmp_enableauthentrapsset;
 int             old_snmp_enableauthentraps;
